Migration to .net 5. Add docker and postgresql.

+++ IHunger/4 - Infra/4.1 - Data/IHunger.Infra.Data/Mappings/ItemMapping.cs
@@ -19,7 +19,7 @@ namespace IHunger.Infra.Data.Mappings
 
             builder.Property(p => p.Quantity)
                .IsRequired()
-               .HasColumnType("TINYINT");
+               .HasColumnType("integer");
         }
     }
 }
