+++ IHunger/4 - Infra/4.1 - Data/IHunger.Infra.Data/Mappings/OrderStatusMapping.cs
@@ -19,7 +19,7 @@ namespace IHunger.Infra.Data.Mappings
 
             builder.Property(p => p.Number)
                 .IsRequired()
-                .HasColumnType("tinyint");
+                .HasColumnType("integer");
         }
     }
 }
